+++ ds4api/callbacks.go
@@ -13,7 +13,6 @@ var (
 	tokenCallbacks            = map[uintptr]TokenEmitFunc{}
 	doneCallbacks             = map[uintptr]GenerationDoneFunc{}
 	progressCallbacks         = map[uintptr]ProgressFunc{}
-	logCallbacks              = map[uintptr]LogFunc{}
 	abortCallbacks            = map[uintptr]AbortFunc{}
 
 	tokenEmitCallback = purego.NewCallback(func(ud uintptr, token int32) {
@@ -39,9 +38,6 @@ var (
 		if fn != nil {
 			fn(goString(event), int(current), int(total))
 		}
-	})
-	logCallback = purego.NewCallback(func(ud uintptr, typ int32, msg unsafe.Pointer) {
-		invokeLogCallback(ud, LogType(typ), goString(msg))
 	})
 	abortCallback = purego.NewCallback(func(ud uintptr, msg unsafe.Pointer) {
 		invokeAbortCallback(ud, goString(msg))
@@ -114,38 +110,6 @@ func unregisterProgressCallback(id uintptr) {
 	callbackMu.Lock()
 	delete(progressCallbacks, id)
 	callbackMu.Unlock()
-}
-
-func registerLogCallback(fn LogFunc) uintptr {
-	if fn == nil {
-		return 0
-	}
-	callbackMu.Lock()
-	defer callbackMu.Unlock()
-	id := nextCallbackID()
-	logCallbacks[id] = fn
-	return id
-}
-
-func unregisterLogCallback(id uintptr) {
-	if id == 0 {
-		return
-	}
-	callbackMu.Lock()
-	delete(logCallbacks, id)
-	callbackMu.Unlock()
-}
-
-func invokeLogCallback(id uintptr, typ LogType, msg string) {
-	if id == 0 {
-		return
-	}
-	callbackMu.Lock()
-	fn := logCallbacks[id]
-	callbackMu.Unlock()
-	if fn != nil {
-		fn(typ, msg)
-	}
 }
 
 func registerAbortCallback(fn AbortFunc) uintptr {