+++ CHANGELOG.md
@@ -1,13 +1,15 @@
 # `ds4go` CHANGELOG
 
-NOTE: This currently needs a patched `ds4` to make a shared library and route logging and aborts; we also embed the `.metal` files.   See https://github.com/neomantra/ds4/tree/nm-shared-stderr
+NOTE: This currently needs a patched `ds4` to make a shared library and route logging and aborts; we also embed the `.metal` files.   See https://github.com/neomantra/ds4/tree/nm-shared
 
 ## v0.5.0 (unreleased)
 
+ * **BREAKING** We changed how logging works in our `nm-shared` branch of `ds4`, so there is now `SetStderrFd` rather than `SetLogFunc`.
  * Add [`lsp` package](./lsp/README.md) for Language Server Protocol (LSP) support
  * Add [`webtool` package](./webtool/README.md) for searching and querying web pages using Google Chrome.
  * Add `ds4go web search` and `ds4go web visit` CLI tools
  * Add `ds4.DetectDefaultBackend`
+ * Add `ds4.CaptureStderr(io.Writer)` helper 
 
 ## v0.4.0 (2026-05-27)